重命名：重命名所有程序

+++ ocrs_03_plt_gray_dev.py
@@ -28,12 +28,8 @@ import matplotlib.pyplot as plt
 
 
 def run(pair, hdf5_file):
-    if os.path.isfile(hdf5_file):
-        log.error("hdf5 file not exist： ".format(hdf5_file))
-        return
-
     # 加载程序配置文件
-    proj_cfg_file = os.path.join(main_path, "cfg", "global.yaml")
+    proj_cfg_file = os.path.join(main_path, "global.yaml")
     proj_cfg = loadYamlCfg(proj_cfg_file)
     if proj_cfg is None:
         log.error("Not find the config file: {}".format(proj_cfg_file))
@@ -43,30 +39,43 @@ def run(pair, hdf5_file):
     try:
         datasets = proj_cfg["plt_gray"][pair].get("datasets")
         filename_suffix = proj_cfg["plt_gray"][pair].get("filename_suffix")
-    except ValueError:
+    except Exception as why:
+        print why
         log.error("Please check the yaml plt_gray args")
         return
 
-    # 开始绘图 ------------------------------------------------------
-    log.info("start plot gray picture")
+    print '-' * 100
+    print "Start plot gray picture."
+    if not os.path.isfile(hdf5_file):
+        log.error("File not exist: {}".format(hdf5_file))
+        return
+
     file_name = os.path.splitext(hdf5_file)[0]
     out_pic = "{}_{}.{}".format(file_name, filename_suffix, "png")
 
-    with h5py.File(hdf5_file, 'r') as h5:
-        if len(datasets) == 3:
-            datas = []
-            for set_name in datasets:
-                datas.append(h5.get(set_name)[:])
-            dv_rgb(datas[0], datas[1], datas[2], out_pic)
-            log.info(out_pic)
-        elif len(datasets) == 1:
-            for set_name in datasets:
-                data = h5.get(set_name)
-                plt.imshow(data, cmap=plt.cm.gray)
-                plt.savefig(out_pic, dpi=200)
-                log.info(out_pic)
-        else:
-            log.error("datasets must be 1 or 3")
+    try:
+        with h5py.File(hdf5_file, 'r') as h5:
+            if len(datasets) == 3:
+                datas = []
+                for set_name in datasets:
+                    datas.append(h5.get(set_name)[:])
+                dv_rgb(datas[0], datas[1], datas[2], out_pic)
+            elif len(datasets) == 1:
+                for set_name in datasets:
+                    data = h5.get(set_name)
+                    plt.imshow(data, cmap=plt.cm.gray)
+                    plt.savefig(out_pic, dpi=200)
+
+            else:
+                log.error("datasets must be 1 or 3")
+                return
+    except Exception as why:
+        print why
+        log.error("Plot gray picture error: {}".format(hdf5_file))
+        return
+
+    print "Output picture: {}".format(out_pic)
+    print '-' * 100
 
 
 ######################### 程序全局入口 ##############################
@@ -85,7 +94,7 @@ if __name__ == "__main__":
     # 获取程序所在位置，拼接配置文件
     main_path, main_file = os.path.split(os.path.realpath(__file__))
     project_path = main_path
-    config_file = os.path.join(project_path, "cfg", "global.cfg")
+    config_file = os.path.join(project_path, "global.cfg")
 
     # 配置不存在预警
     if not os.path.isfile(config_file):
@@ -98,9 +107,9 @@ if __name__ == "__main__":
     log = LogServer(LOG_PATH)
 
     # 开启进程池
-    thread_number = inCfg["CROND"]["threads"]
+    # thread_number = inCfg["CROND"]["threads"]
     # thread_number = 1
-    pool = Pool(processes=int(thread_number))
+    # pool = Pool(processes=int(thread_number))
 
     if not len(args) == 2:
         print help_info
@@ -114,3 +123,5 @@ if __name__ == "__main__":
         file_path = args[1]
         run(sat_sensor, file_path)
         # pool.apply_async(run, (sat_sensor, file_path))
+        # pool.close()
+        # pool.join()
